+++ src/event/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod invoice_created_dto;
+pub mod invoice_dto;
