+++ src/graphql/model/invoice.rs
@@ -1,13 +1,13 @@
-use async_graphql::{Error, SimpleObject};
-use bson::{DateTime, Uuid};
+use async_graphql::{Error, Result, SimpleObject};
+use bson::{doc, DateTime, Uuid};
+use mongodb::{options::FindOneOptions, Collection};
 use serde::{Deserialize, Serialize};
 
-use crate::{
+use crate::event::http_event_service::{HttpEventServiceState, OrderEventData};
+
+use super::{
+    super::query::query_object,
     foreign_types::{User, UserAddress, VendorAddress},
-    http_event_service::{HttpEventServiceState, OrderEventData},
-    query::{
-        project_user_to_user_address, query_object, query_user_address_user, query_vendor_address,
-    },
 };
 
 static INVOICE_TERMS: &str = "This invoice is created according the the companies terms and conditions specified on the website.";
@@ -25,7 +25,7 @@ pub struct Invoice {
 }
 
 impl Invoice {
-    /// Creates a new invoice from OrderEventData and HttpEventServiceState (containing the database connections).
+    /// Creates a new invoice from `OrderEventData` and `HttpEventServiceState` (containing the database connections).
     pub async fn new(
         order_event_data: OrderEventData,
         state: &HttpEventServiceState,
@@ -105,7 +105,7 @@ Total compensatable amount: {}
     }
 }
 
-/// Sets up all the attributes from OrderEventData and HttpEventServiceState (containing the database connections) that are required for invoice creation.
+/// Sets up all the attributes from `OrderEventData` and `HttpEventServiceState` (containing the database connections) that are required for invoice creation.
 async fn invoice_attribute_setup(
     order_event_data: &OrderEventData,
     state: &HttpEventServiceState,
@@ -146,35 +146,49 @@ fn build_order_item_invoice_content(value: &OrderEventData) -> String {
     content
 }
 
-/// DTO of an invoice for an order.
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct InvoiceDTO {
-    pub order_id: Uuid,
-    pub issued_at: chrono::DateTime<chrono::Utc>,
-    pub content: String,
-}
-
-impl From<Invoice> for InvoiceDTO {
-    fn from(value: Invoice) -> Self {
-        InvoiceDTO {
-            order_id: value.order_id,
-            issued_at: value.issued_at.to_chrono(),
-            content: value.content,
-        }
+/// Shared function to query an address from a MongoDB collection of users.
+/// Returns User which only contains the queried address.
+pub async fn query_user_address_user(
+    collection: &mongodb::Collection<User>,
+    address_id: Uuid,
+) -> Result<User> {
+    let find_options = FindOneOptions::builder()
+        .projection(Some(doc! {
+            "addresses.$": 1,
+            "_id": 1
+        }))
+        .build();
+    let message = format!("Address of UUID: `{}` not found.", address_id);
+    match collection
+        .find_one(
+            doc! {"addresses": {
+                "$elemMatch": {
+                    "_id": address_id
+                }
+            }},
+            Some(find_options),
+        )
+        .await
+    {
+        Ok(maybe_user) => maybe_user.ok_or(Error::new(message.clone())),
+        Err(e) => Err(e.into()),
     }
 }
 
-/// DTO which describes the event context on invoice creation.
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct InvoiceCreatedDTO {
-    pub order: OrderEventData,
-    pub invoice: InvoiceDTO,
+/// Projects result of user address query, which is of type `User`, to the contained user address.
+pub fn project_user_to_user_address(user: User) -> Result<UserAddress> {
+    let message = format!("Projection failed, address could not be extracted from user.");
+    user.addresses
+        .iter()
+        .next()
+        .cloned()
+        .ok_or(Error::new(message.clone()))
 }
 
-impl From<(OrderEventData, InvoiceDTO)> for InvoiceCreatedDTO {
-    fn from((order, invoice): (OrderEventData, InvoiceDTO)) -> Self {
-        Self { order, invoice }
-    }
+/// Shared function to query the current vendor address.
+pub async fn query_vendor_address(collection: &Collection<VendorAddress>) -> Result<VendorAddress> {
+    collection
+        .find_one(None, None)
+        .await?
+        .ok_or(Error::new("Vendor address is not set locally."))
 }
