+++ src/graphql/model/mod.rs
@@ -1,0 +1,3 @@
+pub mod foreign_types;
+pub mod invoice;
+pub mod order;
