+++ src/graphql/query.rs
@@ -1,22 +1,19 @@
 use std::any::type_name;
 
-use crate::{
-    foreign_types::{User, UserAddress, VendorAddress},
-    invoice::Invoice,
-    order::Order,
-};
 use async_graphql::{Context, Error, Object, Result};
 
 use bson::Uuid;
-use mongodb::{bson::doc, options::FindOneOptions, Collection, Database};
+use mongodb::{bson::doc, Collection, Database};
 use serde::Deserialize;
+
+use super::model::{invoice::Invoice, order::Order};
 
 /// Describes GraphQL invoice queries.
 pub struct Query;
 
 #[Object]
 impl Query {
-    /// Entity resolver for order of specific id.
+    /// Entity resolver for order of specific UUID.
     #[graphql(entity)]
     async fn order_entity_resolver<'a>(
         &self,
@@ -30,7 +27,7 @@ impl Query {
         Ok(order)
     }
 
-    /// Entity resolver for invoice of specific id.
+    /// Entity resolver for invoice of specific UUID.
     #[graphql(entity)]
     async fn invoice_entity_resolver<'a>(
         &self,
@@ -43,7 +40,7 @@ impl Query {
         Ok(invoice)
     }
 
-    /// Query for invoice of specific id.
+    /// Query for invoice of specific UUID.
     async fn invoice<'a>(
         &self,
         ctx: &Context<'a>,
@@ -56,54 +53,7 @@ impl Query {
     }
 }
 
-/// Shared function to query an address from a MongoDB collection of users.
-/// Returns User which only contains the queried address.
-pub async fn query_user_address_user(
-    collection: &mongodb::Collection<User>,
-    address_id: Uuid,
-) -> Result<User> {
-    let find_options = FindOneOptions::builder()
-        .projection(Some(doc! {
-            "addresses.$": 1,
-            "_id": 1
-        }))
-        .build();
-    let message = format!("Address of UUID: `{}` not found.", address_id);
-    match collection
-        .find_one(
-            doc! {"addresses": {
-                "$elemMatch": {
-                    "_id": address_id
-                }
-            }},
-            Some(find_options),
-        )
-        .await
-    {
-        Ok(maybe_user) => maybe_user.ok_or(Error::new(message.clone())),
-        Err(e) => Err(e.into()),
-    }
-}
-
-/// Projects result of user address query, which is of type User, to the contained UserAddress.
-pub fn project_user_to_user_address(user: User) -> Result<UserAddress> {
-    let message = format!("Projection failed, address could not be extracted from user.");
-    user.addresses
-        .iter()
-        .next()
-        .cloned()
-        .ok_or(Error::new(message.clone()))
-}
-
-/// Shared function to query the current vendor address.
-pub async fn query_vendor_address(collection: &Collection<VendorAddress>) -> Result<VendorAddress> {
-    collection
-        .find_one(None, None)
-        .await?
-        .ok_or(Error::new("Vendor address is not set locally."))
-}
-
-/// Shared function to query an invoice by an order id.
+/// Shared function to query an invoice by an order UUID.
 pub async fn query_invoice_by_order_id(
     collection: &Collection<Invoice>,
     order_id: Uuid,
@@ -115,7 +65,7 @@ pub async fn query_invoice_by_order_id(
         .ok_or(Error::new(message))
 }
 
-/// Shared function to query an object: T from a MongoDB collection of object: T.
+/// Shared function to query an object: `T` from a MongoDB collection of object: `T`.
 ///
 /// * `connection` - MongoDB database connection.
 /// * `id` - UUID of object.
